requirements.txt
pyrogram-navy
tgcrypto
python-dotenv
APScheduler==3.10.4
